Desktop/Sanjay Srivatsav/RagPath/requirements.txt
# Core framework
pathway

# LLM and embeddings (using local Ollama + Sentence Transformers)

# Web UI
streamlit==1.31.0

chromadb==0.4.22

# Git operations
gitpython==3.1.41

# Utilities
python-dotenv==1.0.1
requests==2.31.0
